refactor: article slice
