+++ src/components/Catalog/CatalogArticle/CatalogArticle.jsx
@@ -6,7 +6,7 @@ import {
   selectArticle,
   selectError,
   selectLoading,
-} from 'store/slices/article/slice';
+} from 'store/slices/article';
 import {
   NotFoundArticle,
   ArticleHeader,
