+++ src/components/DropdownEdit/DropdownEdit.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { languageList, editList } from 'utils/constants';
 import { useSelector } from 'react-redux';
-import { selectArticle } from 'store/slices/article/slice';
+import { selectArticle } from 'store/slices/article';
 import './DropdownEdit.css';
 import { getCurrentTheme } from 'store/slices/theme';
