+++ src/components/NewArticle/NewArticle.jsx
@@ -17,7 +17,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import { changeDraft } from 'store/slices';
 import { getCurrentTheme, getIsThemeLight } from 'store/slices/theme';
 import Recommend from 'components/Recommendations/Recommend';
-import { selectTitles } from 'store/slices/article/slice';
+import { selectTitles } from 'store/slices/article';
 import { Link, useLocation } from 'react-router-dom';
 import clsx from 'clsx';
 import useSubCategoryCheck from 'utils/hooks/useSubCategoryCheck';
