+++ src/components/Reviewers/Reviewers.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Link, useParams } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import { selectArticle } from 'store/slices/article/slice';
+import { selectArticle } from 'store/slices/article';
 import avatar from 'images/admin/avatar_default.svg';
 import './Reviewers.css';
 import ModalAuthor from 'components/Author/ModalAuthor';
