+++ src/store/slices/article/index.js
@@ -1,0 +1,2 @@
+export * from './catalog';
+export * from './slice';
