+++ src/store/slices/index.js
@@ -4,5 +4,3 @@ import _userReducer from './userSlice';
 
 export const userReducer = _userReducer;
 export * from './userSlice';
-
-export { default as articleReducer } from './article/slice';
